requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
fastapi==0.115.0
pymongo==4.7.2
pyniryo==1.1.2
python-dotenv==1.0.1